I've found an issue into the power consumption measurement
Basically, i collect watt-hour value from a power meter and i calculate the average every 60 seconds. After that, i apply rules and et cetera. So this average has to be "well constructed" and I've discovered that it was not.
Endeed, during this slot of 60 seconds I collect several values and depending of them, the number of items varies. I have more data with high value and less with low value. It's because power meter sends a new value anytime it counts a new watt. And it counts more watt per second when i consume more than i consume less... (think about led that blinks more often on your power meter when your hot water tank is started). So at the end of my slot time window i have for example, 10 high values for 1 low and my average is "totally" wrong.
So to be more relevant, i now collect watt indices during the same time slot and i calculate the difference between the first and last indice. It should give me the number of watt consumed during 60 seconds with a more precise resolution.