Back module ported into Datastore module
I've removed the Back module written in Java and using a Rule Engine.
The main objective was to reduce the number of services needed and the
footprint on my computer. Ok! it was because it's fun to code in golang.
But I also needed to translate tasmota telemetries `tele/+/SENSOR` to
`sensors/+/temp`, i had to write a `tasmota-converter` script 8-(

So I've introduced a compute element that listens indices from my linky.
To be compatible with the rest of the stack, i had to publish computed
values on MQTT (but it should not necessary anymore).